+++ src/IServiceCollectionExtensions.cs
@@ -1,14 +1,20 @@
-namespace DefaultNamespace
+namespace BetterHostedServices
 {
-    using BetterHostedServices;
     using Microsoft.Extensions.DependencyInjection;
 
+    /// <summary>
+    /// Extensions on IServiceCollection
+    /// </summary>
     public static class IServiceCollectionExtensions
     {
+        /// <summary>
+        /// Add the scaffolding needed to make the BetterHostedServices etc. work.
+        /// Should be called once before making any IHostedServices run.
+        /// </summary>
         public static IServiceCollection AddBetterHostedServices(this IServiceCollection serviceCollection)
         {
             return serviceCollection
-                .AddTransient<IApplicationEnder, ApplicationEnder>();
+                .AddTransient<IApplicationEnder, ApplicationLifetimeEnder>();
         }
     }
 }
