+++ tests/IntegrationUtils/DummyStartup.cs
@@ -1,9 +1,7 @@
-namespace BetterHostedServices.Test
+namespace BetterHostedServices.Test.IntegrationUtils
 {
-    using DefaultNamespace;
     using Microsoft.AspNetCore.Builder;
     using Microsoft.AspNetCore.Hosting;
-    using Microsoft.AspNetCore.Http;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Hosting;
